+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import dart_v2ray
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  DartV2rayPlugin.register(with: registry.registrar(forPlugin: "DartV2rayPlugin"))
 }