+++ macos/Classes/DartV2rayPlugin.h
@@ -1,0 +1,4 @@
+#import <FlutterMacOS/FlutterMacOS.h>
+
+@interface DartV2rayPlugin : NSObject <FlutterPlugin>
+@end