+++ macos/Classes/DesktopV2rayCoreBridge.cc
@@ -1,0 +1,1 @@
+#include "../../shared/desktop_v2ray_core.cc"